requirements.txt
chainlit==0.7.1
chromadb==0.4.13
google-search-results==2.4.2
langchain==0.0.304
openai==0.28.1
python-dotenv==1.0.0
tiktoken==0.5.1
unstructured==0.10.18